utilizing bit chunks
