+++ src/compression.rs
@@ -8,7 +8,34 @@ use halo2_proofs::{
 use group::ff::{Field, PrimeField};
 use std::marker::PhantomData;
 mod compression_gate;
+mod bit_chunk;
 use compression_gate::CompressionGate;
+use bit_chunk::BitChunkSpread;
+
+// BLAKE2 Sigma constant
+pub const BLAKE2B_SIGMA: [[u8; 16]; 12] = [
+    [0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15],
+    [14, 10, 4, 8, 9, 15, 13, 6, 1, 12, 0, 2, 11, 7, 5, 3],
+    [11, 8, 12, 0, 5, 2, 15, 13, 10, 14, 3, 6, 7, 1, 9, 4],
+    [7, 9, 3, 1, 13, 12, 11, 14, 2, 6, 5, 10, 4, 0, 15, 8],
+    [9, 0, 5, 7, 2, 4, 10, 15, 14, 1, 11, 12, 6, 8, 3, 13],
+    [2, 12, 6, 10, 0, 11, 8, 3, 4, 13, 7, 5, 15, 14, 1, 9],
+    [12, 5, 1, 15, 14, 13, 4, 10, 0, 7, 6, 3, 9, 2, 8, 11],
+    [13, 11, 7, 14, 12, 1, 3, 9, 5, 0, 15, 4, 8, 6, 2, 10],
+    [6, 15, 14, 9, 11, 3, 0, 8, 12, 2, 13, 7, 1, 4, 10, 5],
+    [10, 2, 8, 4, 7, 6, 1, 5, 15, 11, 9, 14, 3, 12, 13, 0],
+];
+
+pub const BLAKE2B_IV: [u64; 8] = [
+    0x6a09e667f3bcc908,
+    0xbb67ae8584caa73b,
+    0x3c6ef372fe94f82b,
+    0xa54ff53a5f1d36f1,
+    0x510e527fade682d1,
+    0x9b05688c2b3e6c1f,
+    0x1f83d9abfb41bd6b,
+    0x5be0cd19137e2179,
+];
 
 // This is where we will define the message and state chunks that serve as inputs to the compression function
 #[derive(Clone, Debug)]
@@ -76,7 +103,11 @@ struct CompressionConfig {
 }
 
 impl CompressionConfig {
-    fn configure(meta: &mut ConstraintSystem<pallas::Base>) -> Self {
+    pub(super) fn configure(
+        meta: &mut ConstraintSystem<pallas::Base>,
+        message: [Column<Advice>; 16],
+
+    ) -> Self {
         // define advice and selectors
         Self {
             // define advice and selectors
@@ -92,9 +123,21 @@ impl CompressionConfig {
         x: Expression<F>,
         y: Expression<F>,
     ) -> Vec<Expression<F>> {
-        let r1 = lookup.r1;
-        let r2 = lookup.r2;
-        let r3 = lookup.r3;
+        let r1 = meta.lookup_table_column(r1);
+        meta.lookup(|meta| {
+            let r_1 = meta.query_any(a, Rotation::cur());
+            vec![(r_1, r1)]
+        });
+        let r2 = meta.lookup_table_column(r2);
+        meta.lookup(|meta| {
+            let r_2 = meta.query_any(a, Rotation::cur());
+            vec![(r_2, r2)]
+        });
+        let r3 = meta.lookup_table_column(r3);
+        meta.lookup(|meta| {
+            let r_3 = meta.query_any(a, Rotation::cur());
+            vec![(r_3, r3)]
+        });
         let r4 = meta.lookup_table_column(r4);
         meta.lookup(|meta| {
             let r_4 = meta.query_any(a, Rotation::cur());
