+++ src/compression/compression_gate.rs
@@ -2,6 +2,8 @@ use halo2_proofs::plonk::Expression;
 
 use super::super::utils::*;
 use super::super::compression::{StateChunk, MessageChunk};
+mod bit_chunk;
+use bit_chunk::BitChunkSpread;
 
 pub struct CompressionGate<F: Field>(PhantomData<F>);
 
@@ -11,7 +13,7 @@ impl<F: PrimeField> CompressionGate<F> {
     }
 
     // Implement G function
-    pub fn g_func(Vec<StateChunk>>, a: Expression<F>, b: Expression<F>, c: Expression<F>, d: Expression<F>, x: MessageChunk, y: MessageChunk, r1: Expression<F>, r2: Expression<F>, r3: Expression<F>, r4: Expression<F>) -> Vec<StateChunk>> {
+    pub fn g_func(working_vec: Vec<BitChunkSpread>>, a: Expression<F>, b: Expression<F>, c: Expression<F>, d: Expression<F>, x: MessageChunk, y: MessageChunk, r1: Expression<F>, r2: Expression<F>, r3: Expression<F>, r4: Expression<F>) -> Vec<BitChunkSpread>> {
         let w = 64; // Word size
         let r1 = 32;
         let r2 = 24;
@@ -26,7 +28,8 @@ impl<F: PrimeField> CompressionGate<F> {
         let tmp6 = v[d] ^ tmp5;
         let tmp7 = v[c] + tmp6;
         let tmp8 = v[b] ^ tmp7;
-    
+        
+        // TODO: replace rotate_right with >>> operators
         v[a] = tmp1;
         v[d] = tmp2.rotate_right(r1);
         v[c] = tmp3;
